requirements.txt
adjustText==1.2.0
affine==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.4.0
apispec==6.6.1
apsw==3.46.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==6.1.3
astropy-iers-data==0.2024.8.27.10.28.29
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blinker==1.8.2
boto3==1.34.151
botocore==1.34.151
Bottleneck==1.4.0
branca==0.7.2
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.5.0
cads-api-client==1.2.0
Cartopy==0.23.0
cattrs==24.1.2
cdsapi==0.7.1
celery==5.4.0
certifi==2024.7.4
cf-units==3.2.0
cffi==1.16.0
cfgrib==0.9.14.0
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
conflator==0.1.5
contextily==1.6.2
contourpy==1.2.1
cron-descriptor==1.4.5
croniter==3.0.3
cryptography==42.0.8
cycler==0.12.1
dask==2024.9.1
dask-expr==1.1.11
dask-gateway==2023.9.0
debugpy==1.8.2
decorator==5.1.1
dedl-stack-client==2024.2.2
dedllogin==0.4
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
destinelab==0.4
distributed==2024.9.1
dnspython==2.6.1
docstring_parser==0.16
earthkit-data==0.9.3
earthkit-geo==0.2.0
earthkit-maps==0.0.19
earthkit-meteo==0.1.0
earthkit-regrid==0.3.2
eccodes==1.7.1
eccodes-python==0.9.9
ecmwf-api-client==1.6.3
ecmwflibs==0.6.3
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.112.2
fastjsonschema==2.20.0
filelock==3.15.4
findlibs==0.0.5
Flask==2.3.3
Flask-AppBuilder==4.5.1
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-Compress==1.15
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
folium==0.17.0
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
func-timeout==4.3.5
geographiclib==2.0
geojson==3.1.0
geopandas==1.0.1
geopy==2.4.1
google-auth==2.35.0
greenlet==3.1.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.11.0
hashids==1.3.1
HDA==0.1.0
healpy==1.17.3
holidays==0.25
httpcore==1.0.6
httpx==0.27.2
humanize==4.10.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpath-ng==1.5.3
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kombu==5.4.2
korean-lunar-calendar==0.3.1
limits==3.13.0
llvmlite==0.43.0
locket==1.0.0
lxml==5.2.2
Mako==1.3.5
mapclassify==2.8.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mercantile==1.2.1
mistune==3.0.2
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiurl==0.3.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.1.post1
networkx==3.3
nh3==0.2.18
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.1
numpy==1.26.4
nvidia-nccl-cu12==2.23.4
ordered-set==4.1.0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
OWSLib==0.31.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
paramiko==3.5.0
parsedatetime==2.6
parso==0.8.4
partd==1.4.2
pdbufr==0.11.0
pexpect==4.9.0
pgsanity==0.2.9
pgvector==0.3.5
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
ply==3.11
polars==1.7.1
polyline==2.0.2
polytope-client==0.7.4
prison==0.2.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pvlib==0.11.0
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyerfa==2.0.1.4
Pygments==2.18.0
pygrib==2.1.6
PyJWT==2.9.0
PyNaCl==1.5.0
pyogrio==0.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
PySocks==1.7.1
pystac==1.10.1
pystac-client==0.8.3
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-geohash==0.8.5
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
rasterio==1.4.1
redis==4.6.0
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
requests-ftp==0.3.1
requests-futures==1.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-argparse==1.5.2
rpds-py==0.19.1
rsa==4.9
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.9.1
Send2Trash==1.8.3
shap==0.46.0
shapely==2.0.6
shillelagh==1.2.28
shortid==0.1.2
simplejson==3.19.3
six==1.16.0
slack_sdk==3.33.1
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.38.3
sqlglot==23.17.0
sqlparse==0.4.4
sshtunnel==0.4.0
stack-data==0.6.3
starlette==0.38.2
stream-zip==0.0.82
tabulate==0.8.10
tblib==3.0.0
tenacity==9.0.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.2.2
usgs==0.3.5
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
Whoosh==2.7.4
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
WTForms-JSON==0.3.5
xarray==2024.7.0
xgboost==2.1.1
XlsxWriter==3.0.9
xyzservices==2024.9.0
yarl==1.9.4
zict==3.0.0
zipp==3.19.2
zstandard==0.23.0